refactor(plugin-reporter): absorb thymian: report claim, drop plugin-thymian-report
Rescind @thymian/plugin-thymian-report per the 2026-08-12 course
correction (thymian-internal#507 rework R1/R2/R4): @thymian/plugin-reporter
owns persisted-report I/O in both directions and now claims thymian:
inputs on core.report.convert.

- relocate load-thymian-report.ts verbatim (same-package formatter
  wording in the header; error messages byte-identical)
- register the identity-preserving claim/convert handler via
  src/thymian-report-input.ts inside the existing plugin() body; add
  'core.report.convert' to actions.listensOn
- relocate all 12 tests (runtime style kept; reporterPlugin registered
  with empty formatters); disableConsoleIntercept parity in
  vitest.config.ts
- delete packages/plugin-thymian-report and unwind wiring: root +
  thymian tsconfig references, thymian dependency, commitlint scope,
  license-checker entry, default-config autoload, type-only import in
  report merge (now pointing at the claiming plugin), e2e fixture
  plugin entry, lockfile
- refresh stale tool-name fixture strings in merge.test.ts

Behavior-identical: merge semantics, error messages, and exit codes
unchanged; 85/85 e2e green.
