+++ .license-checker.json
@@ -26,7 +26,6 @@
     "@thymian/rules-api-description-validation",
     "@thymian/rules-rfc-9110",
     "@thymian/plugin-har",
-    "@thymian/plugin-spectral",
-    "@thymian/plugin-thymian-report"
+    "@thymian/plugin-spectral"
   ]
 }
