+++ commitlint.config.js
@@ -26,7 +26,6 @@ export default {
         'thymian',
         'e2e',
         'plugin-har',
-        'plugin-thymian-report',
       ],
     ],
     'scope-empty': [2, 'never'],
