+++ packages/common-cli/src/default-config.ts
@@ -23,6 +23,5 @@ export const defaultConfig: ThymianConfig = {
     '@thymian/plugin-http-analyzer': {},
     '@thymian/plugin-har': {},
     '@thymian/plugin-spectral': {},
-    '@thymian/plugin-thymian-report': {},
   },
 };
