+++ packages/plugin-reporter/vitest.config.ts
@@ -3,5 +3,6 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     include: ['test/**/*.test.ts'],
+    disableConsoleIntercept: true,
   },
 });
