--- packages/plugin-thymian-report/project.json
@@ -1,16 +1,0 @@
-{
-  "name": "plugin-thymian-report",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/plugin-thymian-report/src",
-  "projectType": "library",
-  "tags": ["scope:plugin", "type:lib", "npm:public"],
-  "targets": {
-    "build": {
-      "dependsOn": ["^build"],
-      "command": "tsc --build tsconfig.lib.json",
-      "options": {
-        "cwd": "packages/plugin-thymian-report"
-      }
-    }
-  }
-}
