--- packages/plugin-thymian-report/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "files": [],
-  "include": ["src/**/*.ts", "test/**/*.ts"],
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "dist"
-  },
-  "references": [
-    {
-      "path": "../core"
-    },
-    {
-      "path": "./tsconfig.lib.json"
-    }
-  ],
-  "nx": {
-    "addTypecheckTarget": false
-  }
-}
