+++ packages/thymian/package.json
@@ -62,7 +62,6 @@
     "@thymian/plugin-request-dispatcher": "0.0.0-PLACEHOLDER",
     "@thymian/plugin-sampler": "0.0.0-PLACEHOLDER",
     "@thymian/plugin-spectral": "0.0.0-PLACEHOLDER",
-    "@thymian/plugin-thymian-report": "0.0.0-PLACEHOLDER",
     "@thymian/plugin-websocket-proxy": "0.0.0-PLACEHOLDER",
     "@thymian/rules-rfc-9110": "0.0.0-PLACEHOLDER",
     "@thymian/rules-api-description-validation": "0.0.0-PLACEHOLDER",
