+++ tsconfig.json
@@ -56,9 +56,6 @@
     },
     {
       "path": "./packages/plugin-har"
-    },
-    {
-      "path": "./packages/plugin-thymian-report"
     }
   ]
 }
